+++ src/config-un/config.ts
@@ -199,7 +199,6 @@ export const eslintConfigInternal = async <const ExtraPlugins extends ExtraPlugi
     extraConfigs,
     extraPlugins,
     ignores,
-    overrideIgnores,
     pluginRenames = {},
     loadPluginsOnDemand,
     disablePrettierIncompatibleRules,
@@ -408,8 +407,10 @@ export const eslintConfigInternal = async <const ExtraPlugins extends ExtraPlugi
     .disableAnyRule('import', 'no-default-export');
 
   // According to ESLint docs: "If `ignores` is used without any other keys in the configuration object, then the patterns act as global ignores <...> Patterns are added after the default patterns, which are ["**/node_modules/", ".git/"]." - https://eslint.org/docs/latest/use/configure/configuration-files#globally-ignoring-files-with-ignores
-  const globalIgnores = [...(overrideIgnores ? [] : DEFAULT_GLOBAL_IGNORES), ...(ignores || [])];
-
+  const globalIgnores = [
+    ...(!Array.isArray(ignores) && ignores?.override ? [] : DEFAULT_GLOBAL_IGNORES),
+    ...(Array.isArray(ignores) ? ignores : []),
+  ];
   debug(`Globally ignored files: ${JSON.stringify(globalIgnores)}`);
 
   type UnresolvedConfigType =